refactor(privmx-endpoint/ios): move Utils to core module
