+++ privmx-endpoint/src/iosMain/kotlin/com/simplito/kotlin/privmx_endpoint/modules/core/utils/Base32.ios.kt
@@ -1,4 +1,4 @@
-package com.simplito.java.privmx_endpoint.model.utils
+package com.simplito.kotlin.privmx_endpoint.modules.core.utils
 
 import cnames.structs.pson_value
 import com.simplito.kotlin.privmx_endpoint.utils.asResponse
@@ -15,12 +15,12 @@ import libprivmxendpoint.pson_free_result
 import libprivmxendpoint.pson_free_value
 
 @OptIn(ExperimentalForeignApi::class)
-actual object Base64 {
+actual object Base32 {
     actual fun encode(data: ByteArray): String = memScoped {
         val pson_result = allocPointerTo<pson_value>()
         val args = makeArgs(data.pson)
         try {
-            privmx_endpoint_execUtils(_nativeUtils.value, 7, args, pson_result.ptr)
+            privmx_endpoint_execUtils(_nativeUtils.value, 4, args, pson_result.ptr)
             pson_result.value?.asResponse?.getResultOrThrow()?.typedValue()!!
         } finally {
             pson_free_value(args)
@@ -28,11 +28,11 @@ actual object Base64 {
         }
     }
 
-    actual fun decode(base64Data: String): ByteArray = memScoped {
+    actual fun decode(base32Data: String): ByteArray = memScoped {
         val pson_result = allocPointerTo<pson_value>()
-        val args = makeArgs(base64Data.pson)
+        val args = makeArgs(base32Data.pson)
         try {
-            privmx_endpoint_execUtils(_nativeUtils.value, 8, args, pson_result.ptr)
+            privmx_endpoint_execUtils(_nativeUtils.value, 5, args, pson_result.ptr)
             pson_result.value?.asResponse?.getResultOrThrow()?.typedValue()!!
         } finally {
             pson_free_value(args)
@@ -44,7 +44,7 @@ actual object Base64 {
         val pson_result = allocPointerTo<pson_value>()
         val args = makeArgs(data.pson)
         try {
-            privmx_endpoint_execUtils(_nativeUtils.value, 9, args, pson_result.ptr)
+            privmx_endpoint_execUtils(_nativeUtils.value, 6, args, pson_result.ptr)
             pson_result.value!!.asResponse?.getResultOrThrow()?.typedValue()!!
         } finally {
             pson_free_value(args)
