+++ privmx-endpoint/src/iosMain/kotlin/com/simplito/kotlin/privmx_endpoint/modules/core/utils/Hex.ios.kt
@@ -1,4 +1,4 @@
-package com.simplito.java.privmx_endpoint.model.utils
+package com.simplito.kotlin.privmx_endpoint.modules.core.utils
 
 import cnames.structs.pson_value
 import com.simplito.kotlin.privmx_endpoint.utils.asResponse
